requirements.txt
Django==4.1.5
django-debug-toolbar==3.8.1
Pillow==9.4.0
environs[django]==9.3.2
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
django-filter==22.1
Markdown==3.4.1
requests==2.28.2
geopy==2.3.0
phonenumbers==8.13.4
gunicorn==20.1.0
rollbar==0.16.3
GitPython==3.1.31

